2.11: The Phonebook Step6
Store the initial state of the application in the file db.json. Start json-server on port 3001 and make sure that the server returns the list of people by going to the address http://localhost:3001/persons in the browser. Modify the application such that the initial state of the data is fetched from the server using the axios-library.